+++ src/TetMesh.h
@@ -91,6 +91,7 @@ public:
          */
         bool is_valid(const TetMesh& m) const;
         bool is_boundary_edge(const TetMesh& m) const;
+        bool is_boundary_face(const TetMesh& m) const;
 
 
         void update_version_number(const TetMesh& m);
@@ -277,7 +278,7 @@ public:
     bool split_edge(const Tuple& t, std::vector<Tuple>& new_edges);
     bool collapse_edge(const Tuple& t, std::vector<Tuple>& new_edges);
     bool swap_edge(const Tuple& t);
-    bool smooth_vertex(const Tuple& t);
+    bool swap_face(const Tuple& t);
 
     void
     compact(); // cleans up the deleted vertices or tetrahedra, and fixes the corresponding indices
